Refactor script
Rewrite Markdown for loop so that formatted lines are appended to an empty list, rather than immediately written to file. This allows for better handling of whitespaces: Ensuring that all sections are separated by two whitespaces (for readability when opening the Markdown file directly). Importantly, it ensures that the final Markdown document contains only one trailing whitespace. The previous for loop occasionally led to 2 whitespaces at the end of the file, leading to issues when rendering.